Refactor

+++ src/main/java/com/test/prices/core/domain/PricesRepository.java
@@ -6,5 +6,5 @@ import java.util.Date;
 import java.util.List;
 
 public interface PricesRepository {
-    List<PriceItem> findByDate(int brandId, int productId, Date date);
+    List<PriceItem> findByDate(Long brandId, Long productId, Date date);
 }
