+++ src/main/java/com/test/prices/core/infrastructure/JPAPricesRepositoryImpl.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 public interface JPAPricesRepositoryImpl extends CrudRepository<PriceItem, Long>, PricesRepository {
     @Query("SELECT p FROM PriceItem p WHERE p.brandId = ?1 and p.productId = ?2 and p.startDate <= ?3 and p.endDate >= ?3 order by p.priority desc")
-    List<PriceItem> findByDate(int brandId, int productId, Date date);
+    List<PriceItem> findByDate(Long brandId, Long productId, Date date);
 }
